refactor: replace IconGradient type with GradientState in relevant components and services
